add styling to sign up page, and put all related routes into a file
